backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
chromadb>=0.4.18
PyPDF2>=3.0.0
python-docx>=1.1.0
openai>=1.3.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0

